kbkdf: split off unit tests
